Split up dataset details widgets into separate reusable parts
