+++ koordinates/gui/detail_widgets/__init__.py
@@ -1,0 +1,10 @@
+"""
+Reusable widgets for use in dataset details dialogs
+"""
+
+from .attachment_widget import AttachmentWidget  # NOQA
+from .details_table_widget import DetailsTable  # NOQA
+from .header_widget import HeaderWidget  # NOQA
+from .horizontal_line_widget import HorizontalLine  # NOQA
+from .metadata_widget import MetadataWidget  # NOQA
+from .statistic_widget import StatisticWidget  # NOQA
